:recycle: Move RefreshTableButton to common component.
Why? Better re-use in non travel authorization model components.
